requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
openpyxl>=3.1.2
pandas>=2.1.0
python-dotenv>=1.0.0
pyinstaller>=6.3.0
